requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
python-keycloak==0.26.1